Move children components to separate folder and rename Videos component
