+++ src/components/VideoItem/_children/VideoDetails.jsx
