+++ src/components/VideoItem/_children/VideoThumbnail.jsx
