+++ src/components/VideoItem/index.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import VideoDetails from "./VideoDetails";
-import VideoThumbnail from "./VideoThumbnail";
+import VideoDetails from "./_children/VideoDetails";
+import VideoThumbnail from "./_children/VideoThumbnail";
 
 const StyledItem = styled.div`
   display: flex;
