+++ src/components/VideoList.test.jsx
@@ -1,13 +1,14 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import Videos from "./Videos";
+import VideoList from "./VideoList";
 import data from "../mock/youtube-videos-mock.json";
 
-describe("Videos", () => {
+describe("VideoList", () => {
   test("array with at least one video item is in the document", () => {
-    render(<Videos data={data} />);
+    render(<VideoList data={data} />);
     const isArrayOfVideoItems = Array.isArray(screen.queryAllByRole("article"));
     const isNotEmpty = screen.queryAllByRole("article").length > 0;
-    expect(isArrayOfVideoItems && isNotEmpty).toBe(true);
+    expect(isArrayOfVideoItems).toBe(true);
+    expect(isNotEmpty).toBe(true);
   });
 });
